+++ dragonstone/models/points.py
@@ -1,5 +1,6 @@
 from datetime import datetime
 
+from constance import config
 from django.db import models
 from polymorphic.models import PolymorphicModel
 
@@ -12,7 +13,6 @@ from dragonstone import (
 )
 from dragonstone import managers
 from main import EASY, MEDIUM, HARD, VERY_HARD
-from main.models import Settings
 
 __all__ = [
     "DragonstonePoints",
@@ -60,7 +60,7 @@ class RecruitmentPoints(DragonstonePoints):
 
     def save(self, update_fields=None, *args, **kwargs):
         if not self.pk:
-            self.points = int(Settings.objects.get(name="RECRUITER_PTS"))
+            self.points = config.RECRUITER_PTS
         super().save(*args, **kwargs)
 
 
@@ -74,11 +74,11 @@ class SotMPoints(DragonstonePoints):
     def save(self, update_fields=None, *args, **kwargs):
         if not self.pk:
             if self.rank == 1:
-                self.points = int(Settings.objects.get(name="SOTM_FIRST_PTS"))
+                self.points = config.SOTM_FIRST_PTS
             elif self.rank == 2:
-                self.points = int(Settings.objects.get(name="SOTM_SECOND_PTS"))
+                self.points = config.SOTM_SECOND_PTS
             elif self.rank == 3:
-                self.points = int(Settings.objects.get(name="SOTM_THIRD_PTS"))
+                self.points = config.SOTM_THIRD_PTS
         super().save(*args, **kwargs)
 
 
@@ -95,21 +95,13 @@ class PVMSplitPoints(DragonstonePoints):
     def save(self, update_fields=None, *args, **kwargs):
         if not self.pk:
             if self.submission.content.difficulty == EASY:
-                self.points = int(
-                    Settings.objects.get(name="PVM_SPLITS_EASY_PTS").value
-                )
+                self.points = int(config.PVM_SPLITS_EASY_PTS)
             elif self.submission.content.difficulty == MEDIUM:
-                self.points = int(
-                    Settings.objects.get(name="PVM_SPLITS_MEDIUM_PTS").value
-                )
+                self.points = int(config.PVM_SPLITS_MEDIUM_PTS)
             elif self.submission.content.difficulty == HARD:
-                self.points = int(
-                    Settings.objects.get(name="PVM_SPLITS_HARD_PTS").value
-                )
+                self.points = int(config.PVM_SPLITS_HARD_PTS)
             elif self.submission.content.difficulty == VERY_HARD:
-                self.points = int(
-                    Settings.objects.get(name="PVM_SPLITS_VERY_HARD_PTS").value
-                )
+                self.points = int(config.PVM_SPLITS_VERY_HARD_PTS)
             self.date = self.submission.date
         super().save(*args, **kwargs)
 
@@ -127,15 +119,13 @@ class MentorPoints(DragonstonePoints):
     def save(self, update_fields=None, *args, **kwargs):
         if not self.pk:
             if self.submission.content.difficulty == EASY:
-                self.points = int(Settings.objects.get(name="MENTOR_EASY_PTS").value)
+                self.points = config.MENTOR_EASY_PTS
             elif self.submission.content.difficulty == MEDIUM:
-                self.points = int(Settings.objects.get(name="MENTOR_MEDIUM_PTS").value)
+                self.points = config.MENTOR_MEDIUM_PTS
             elif self.submission.content.difficulty == HARD:
-                self.points = int(Settings.objects.get(name="MENTOR_HARD_PTS").value)
+                self.points = config.MENTOR_HARD_PTS
             elif self.submission.content.difficulty == VERY_HARD:
-                self.points = int(
-                    Settings.objects.get(name="MENTOR_VERY_HARD_PTS").value
-                )
+                self.points = int(config.MENTOR_VERY_HARD_PTS)
             self.date = self.submission.date
         super().save(*args, **kwargs)
 
@@ -153,21 +143,13 @@ class EventHostPoints(DragonstonePoints):
     def save(self, update_fields=None, *args, **kwargs):
         if not self.pk:
             if self.submission.type == PVM or self.submission.type == SKILLING:
-                self.points = int(
-                    Settings.objects.get(name="EVENT_MINOR_HOSTS_PTS").value
-                )
+                self.points = int()
             elif self.submission.type == EVENT_MENTOR:
-                self.points = int(
-                    Settings.objects.get(name="EVENT_MENTOR_HOSTS_PTS").value
-                )
+                self.points = int(config.EVENT_MENTOR_HOSTS_PTS)
             elif self.submission.type == MAJOR:
-                self.points = int(
-                    Settings.objects.get(name="EVENT_MAJOR_HOSTS_PTS").value
-                )
+                self.points = int(config.EVENT_MAJOR_HOSTS_PTS)
             elif self.submission.type == OTHER:
-                self.points = int(
-                    Settings.objects.get(name="EVENT_OTHER_HOSTS_PTS").value
-                )
+                self.points = int(config.EVENT_OTHER_HOSTS_PTS)
         self.date = self.submission.date
         super().save(*args, **kwargs)
 
@@ -185,21 +167,13 @@ class EventParticipantPoints(DragonstonePoints):
     def save(self, update_fields=None, *args, **kwargs):
         if not self.pk:
             if self.submission.type == PVM or self.submission.type == SKILLING:
-                self.points = int(
-                    Settings.objects.get(name="EVENT_MINOR_PARTICIPANTS_PTS").value
-                )
+                self.points = int(config.EVENT_MINOR_PARTICIPANTS_PTS)
             elif self.submission.type == EVENT_MENTOR:
-                self.points = int(
-                    Settings.objects.get(name="EVENT_MENTOR_PARTICIPANTS_PTS").value
-                )
+                self.points = int(config.EVENT_MENTOR_PARTICIPANTS_PTS)
             elif self.submission.type == MAJOR:
-                self.points = int(
-                    Settings.objects.get(name="EVENT_MAJOR_PARTICIPANTS_PTS").value
-                )
+                self.points = int(config.EVENT_MAJOR_PARTICIPANTS_PTS)
             elif self.submission.type == OTHER:
-                self.points = int(
-                    Settings.objects.get(name="EVENT_OTHER_PARTICIPANTS_PTS").value
-                )
+                self.points = int(config.EVENT_OTHER_PARTICIPANTS_PTS)
         self.date = self.submission.date
         super().save(*args, **kwargs)
 
@@ -217,20 +191,12 @@ class EventDonorPoints(DragonstonePoints):
     def save(self, update_fields=None, *args, **kwargs):
         if not self.pk:
             if self.submission.type == PVM or self.submission.type == SKILLING:
-                self.points = int(
-                    Settings.objects.get(name="EVENT_MINOR_DONORS_PTS").value
-                )
+                self.points = int(config.EVENT_MINOR_DONORS_PTS)
             elif self.submission.type == EVENT_MENTOR:
-                self.points = int(
-                    Settings.objects.get(name="EVENT_MENTOR_DONORS_PTS").value
-                )
+                self.points = int(config.EVENT_MENTOR_DONORS_PTS)
             elif self.submission.type == MAJOR:
-                self.points = int(
-                    Settings.objects.get(name="EVENT_MAJOR_DONORS_PTS").value
-                )
+                self.points = int(config.EVENT_MAJOR_DONORS_PTS)
             elif self.submission.type == OTHER:
-                self.points = int(
-                    Settings.objects.get(name="EVENT_OTHER_DONORS_PTS").value
-                )
+                self.points = int(config.EVENT_OTHER_DONORS_PTS)
             self.date = self.submission.date
         super().save(*args, **kwargs)